requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.1
black==23.3.0
blis==0.7.9
boto3==1.26.124
botocore==1.29.124
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
coverage==7.2.3
crytic-compile==0.2.4
cssselect==1.2.0
cymem==2.0.7
distlib==0.3.5
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.8.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
orjson==3.8.10
outcome==1.2.0
packaging==23.0
panda==0.3.1
pandas==2.0.1
pandas-stubs==2.0.0.230412
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==2.5.2
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
preshed==3.0.8
prettytable==3.4.1
protobuf==4.22.1
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pysha3==1.0.2
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
selenium==4.8.3
six==1.16.0
slither-analyzer==0.9.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typer==0.7.0
types-pytz==2023.3.0.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.16.3
wasabi==1.1.1
wcwidth==0.2.5
webdriver-manager==3.8.6
websocket-client==1.5.1
wsproto==1.2.0
yarl==1.8.2
